Fix warnings-as-errors build
CMake: exclude JUCE headers and sources from project warnings
- Mark all JUCE targets as SYSTEM so their include directories use
  -isystem, suppressing warnings in JUCE headers included by our code
- Add deferred function that suppresses warnings on JUCE source files
  compiled inside project targets; handles both PRIVATE sources (e.g.
  vst3_helper) and INTERFACE_SOURCES propagated from module targets

Fix sign-conversion and float-conversion warnings in project code
- SelfModEngine.cpp: cast num_channels to size_t in assign()
- EngineTest.cpp: use size_t loop index for vector access
- TransferCurveTest.cpp: use 0.1f literal to avoid double->float narrowing

+++ src/selfmod_engine/SelfModEngine.cpp
@@ -45,7 +45,7 @@ struct SelfModEngineImpl : public SelfModEngine {
     {
         sample_rate = sample_rate_arg;
         update_filters();
-        channel_states.assign(num_channels, {});
+        channel_states.assign(static_cast<size_t>(num_channels), {});
     }
 
     void process_block(std::span<float* const> channel_data, int num_samples) override
